Move util functions into new service
